spotify-taste-summarizer/pyproject.toml
[project]
name = "spotify-taste-summarizer"
version = "0.1.0"
description = "Summarize your Spotify musical taste into text using AI"
requires-python = ">=3.11"
dependencies = [
    "spotipy>=2.24.0",
    "openai>=1.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/summarizer"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
